+++ src/JK.Common.Tests/Extensions/StringExtensionTests/ToNullableDecimal.cs
@@ -9,7 +9,7 @@ public class ToNullableDecimal
     [InlineData("")]
     [InlineData("  ")]
     [InlineData("abc")]
-    public void ToNullableDecimal_EmptyOrAlpha_Null(string value)
+    public void EmptyOrAlpha_Null(string value)
     {
         var actual = value.ToNullableDecimal();
         Assert.Null(actual);
@@ -18,7 +18,7 @@ public class ToNullableDecimal
     [Theory]
     [InlineData("123", 123)]
     [InlineData("123.456", 123.456)]
-    public void ToNullableDecimal_Numeric_Decimal(string input, decimal expected)
+    public void Numeric_Decimal(string input, decimal expected)
     {
         var actual = input.ToNullableDecimal();
         Assert.Equal(expected, actual);