+++ src/JK.Common.Tests/Extensions/StringExtensionTests/ToNullableInteger.cs
@@ -10,14 +10,14 @@ public class ToNullableInteger
     [InlineData("  ")]
     [InlineData("abc")]
     [InlineData("123.456")]
-    public void ToNullableInteger_EmptyAlphaOrDecimal_Null(string value)
+    public void EmptyAlphaOrDecimal_Null(string value)
     {
         var actual = value.ToNullableInteger();
         Assert.Null(actual);
     }
 
     [Fact]
-    public void ToNullableInteger_Numeric_Int()
+    public void Numeric_Int()
     {
         var actual = "123".ToNullableInteger();
         Assert.Equal(123, actual.Value);